make-good-again

+++ filesystem/iso9660/eltorito.go
@@ -76,8 +76,8 @@ type ElToritoEntry struct {
 	SystemType mbr.Type
 	// LoadSize how many blocks of BootFile to load, equivalent to genisoimage option `-boot-load-size`
 	LoadSize uint16
-	//good: size     uint16
-	size     uint32
+	size     uint16
+	//bad: size     uint32
 	location uint32
 }
 
@@ -128,8 +128,8 @@ func (e *ElToritoEntry) headerBytes(last bool, entries uint16) []byte {
 func (e *ElToritoEntry) entryBytes() []byte {
 	blocks := e.LoadSize
 	if blocks == 0 {
-		//good: blocks = e.size / 512
-		blocks = uint16(e.size / 512)
+		blocks = e.size / 512
+		//bad: blocks = uint16(e.size / 512)
 		if e.size%512 > 1 {
 			blocks++
 		}
@@ -156,8 +156,8 @@ func (e *ElToritoEntry) generateBootTable(pvdSector uint32, p string) ([]byte, e
 	b := make([]byte, 56)
 	binary.LittleEndian.PutUint32(b[0:4], pvdSector)
 	binary.LittleEndian.PutUint32(b[4:8], e.location)
-	//good: binary.LittleEndian.PutUint32(b[8:12], uint32(e.size))
-	binary.LittleEndian.PutUint32(b[8:12], e.size)
+	binary.LittleEndian.PutUint32(b[8:12], uint32(e.size))
+	//bad: binary.LittleEndian.PutUint32(b[8:12], e.size)
 	// Checksum - simply add up all 32-bit words beginning at byte position 64
 	f, err := os.Open(p)
 	if err != nil {
